Move all vending box logic to nyzul folder
